♻️refactor: day off adjustment  schedule mapper
